requirements.txt
fastapi
uvicorn
pydantic
pytest
sqlalchemy
python-multipart
passlib[bcrypt]
python-jose